+++ tests/core.scm
@@ -756,6 +756,23 @@
         (t.is (try (eval '(+ x x)) (catch (e) e.message))
               "Unbound variable `x'")))
 
+(test.only "core: quoted list mutation"
+      (lambda (t)
+        (let ((list '(1 2 3 4)))
+          (set-car! list 10)
+          (t.is list '(10 2 3 4)))))
+
+(test.only "core: freeze list"
+      (lambda (t)
+        (let ((lst '(1 2 3 4)))
+          (lst.freeze)
+          (let loop ((lst lst))
+            (when (not (null? lst))
+              (let ((item (car lst)))
+                (t.is (vector item (to.throw (set-car! lst 10)))
+                      (vector item true)))
+              (loop (cdr list)))))))
+
 ;; TODO
 ;; begin*
 ;; set-obj! throws with null or boolean